src/main/java/org/example/jenkins/EtudiantController.java
package org.example.jenkins;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/etudiants")
public class EtudiantController {
    private final EtudiantService etudiantService;

    @Autowired
    public EtudiantController(EtudiantService etudiantService) {
        this.etudiantService = etudiantService;
    }

    @PostMapping
    public ResponseEntity<EtudiantService.Etudiant> ajouterEtudiant(@RequestParam String nom, @RequestParam String prenom) {
        EtudiantService.Etudiant nouvelEtudiant = etudiantService.ajouterEtudiant(nom, prenom);
        return new ResponseEntity<>(nouvelEtudiant, HttpStatus.CREATED);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> supprimerEtudiant(@PathVariable Long id) {
        etudiantService.supprimerEtudiant(id);
        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
    }

    @GetMapping
    public ResponseEntity<List<EtudiantService.Etudiant>> getAllEtudiants() {
        List<EtudiantService.Etudiant> etudiants = etudiantService.getAllEtudiants();
        return new ResponseEntity<>(etudiants, HttpStatus.OK);
    }
}

